Improve seg tree

+++ content/data-structures/LazySegmentTree.h
@@ -14,26 +14,26 @@
 
 #include "../various/BumpAllocator.h"
 
-const int inf = 1e9;
+const ll inf = 1e9;
 struct Node {
 	Node *l = 0, *r = 0;
-	int lo, hi, mset = inf, madd = 0, val = -inf;
-	Node(int lo,int hi):lo(lo),hi(hi){} // Large interval of -inf
-	Node(vi& v, int lo, int hi) : lo(lo), hi(hi) {
+	ll lo, hi, mset = inf, madd = 0, val = -inf;
+	Node(ll lo,ll hi):lo(lo),hi(hi){} // Large interval of -inf
+	Node(vi& v, ll lo, ll hi) : lo(lo), hi(hi) {
 		if (lo + 1 < hi) {
-			int mid = lo + (hi - lo)/2;
+			ll mid = lo + (hi - lo)/2;
 			l = new Node(v, lo, mid); r = new Node(v, mid, hi);
 			val = max(l->val, r->val);
 		}
 		else val = v[lo];
 	}
-	int query(int L, int R) {
+	ll query(ll L, ll R) {
 		if (R <= lo || hi <= L) return -inf;
 		if (L <= lo && hi <= R) return val;
 		push();
 		return max(l->query(L, R), r->query(L, R));
 	}
-	void set(int L, int R, int x) {
+	void set(ll L, ll R, ll x) {
 		if (R <= lo || hi <= L) return;
 		if (L <= lo && hi <= R) mset = val = x, madd = 0;
 		else {
@@ -41,7 +41,7 @@ struct Node {
 			val = max(l->val, r->val);
 		}
 	}
-	void add(int L, int R, int x) {
+	void add(ll L, ll R, ll x) {
 		if (R <= lo || hi <= L) return;
 		if (L <= lo && hi <= R) {
 			if (mset != inf) mset += x;
@@ -55,7 +55,7 @@ struct Node {
 	}
 	void push() {
 		if (!l) {
-			int mid = lo + (hi - lo)/2;
+			ll mid = lo + (hi - lo)/2;
 			l = new Node(lo, mid); r = new Node(mid, hi);
 		}
 		if (mset != inf)
